% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{batcher}
\alias{batcher}
\title{Batcher (for vectors)}
\usage{
batcher(X, FUN, batch_size, ...)
}
\arguments{
\item{X}{Data to input}

\item{FUN}{Function to apply}

\item{batch_size}{Size of vatches}

\item{...}{Additional arguments to FUN (not yet implemented)}
}
\description{
A function to batch another (simple) function that takes a vector.
}
\examples{
batcher()
}
\keyword{efficiency}
