% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{batcher_array}
\alias{batcher_array}
\title{Batcher (for arrays)}
\usage{
batcher_array(X, FUN, batch_size, ...)
}
\arguments{
\item{X}{Data to input}

\item{FUN}{Function to apply}

\item{batch_size}{Size of vatches}

\item{...}{Additional arguments to FUN (not yet implemented)}
}
\description{
A function to batch another (simple) function that takes an array.
}
\examples{
batcher_array()
}
\keyword{efficiency}
