% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{corner}
\alias{corner}
\title{View Corner of a Matrix}
\usage{
corner(x, min = c(0, 0), max = c(10, 10))
}
\arguments{
\item{x}{Matrix to view}

\item{min}{Minimum row and column indices to view.  Defaults to c(0,0)}

\item{max}{Maximum row and column indices to view.  Defaults to c(10,10)}
}
\description{
This function does for a matrix what head() does for a data frame.
}
\examples{
plotTermCoefs()
}
\keyword{classification}
